requirements.txt
requests
jupyterlab
pyspark==3.2.1
PyYAML
ipykernel
boto3
dnspython
pymongo[srv]
-e .